feat: chronos-portal,update

+++ mail-chronos-portal/src/main/java/com/netease/mail/chronos/portal/controller/RemindTaskManageController.java
@@ -1,10 +1,13 @@
 package com.netease.mail.chronos.portal.controller;
 
+import com.netease.mail.chronos.base.enums.BaseStatusEnum;
+import com.netease.mail.chronos.base.exception.BaseException;
 import com.netease.mail.chronos.base.response.BaseResponse;
 import com.netease.mail.chronos.portal.param.RemindTask;
 import com.netease.mail.chronos.portal.service.SpRemindTaskManageService;
 import com.netease.mail.chronos.portal.vo.RemindTaskVo;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -35,7 +38,10 @@ public class RemindTaskManageController {
      * 删除 任务
      */
     @DeleteMapping
-    public BaseResponse<List<RemindTaskVo>> delete(@RequestParam String colId, @RequestParam(required = false) String compId) {
+    public BaseResponse<List<RemindTaskVo>> delete(@RequestParam(required = false) String colId, @RequestParam(required = false) String compId) {
+        if (StringUtils.isBlank(colId) && StringUtils.isBlank(compId)) {
+            throw new BaseException(BaseStatusEnum.ILLEGAL_ARGUMENT.getCode(), "colId 和 compId 不能都为空");
+        }
         return BaseResponse.success(spRemindTaskManageService.delete(colId, compId));
     }
 
@@ -51,7 +57,10 @@ public class RemindTaskManageController {
      * 查询任务信息
      */
     @GetMapping
-    public BaseResponse<List<RemindTaskVo>> query(@RequestParam String colId, @RequestParam(required = false) String compId) {
+    public BaseResponse<List<RemindTaskVo>> query(@RequestParam(required = false) String colId, @RequestParam(required = false) String compId) {
+        if (StringUtils.isBlank(colId) && StringUtils.isBlank(compId)) {
+            throw new BaseException(BaseStatusEnum.ILLEGAL_ARGUMENT.getCode(), "colId 和 compId 不能都为空");
+        }
         return BaseResponse.success(spRemindTaskManageService.query(colId, compId));
     }
